requirements.txt
# Core Requirements
Django==4.2.21
django-filter==23.2
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-bootstrap5==23.3
django-tables2==2.5.3
django-widget-tweaks==1.4.12

# Database
pyodbc==4.0.39
mssql-django==1.4.1
django-mssql-backend==2.8.1

# Authentication
django-allauth==0.54.0

# API
djangorestframework==3.15.2
django-cors-headers==4.1.0
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.0

# AI Integration
google-generativeai==0.8.3
python-dotenv==1.0.1

# UI Components
django-select2==8.1.2
django-ckeditor==6.5.1
django-jazzmin==2.6.0

# Date/Time
python-dateutil==2.8.2
django-hijri-date==0.1.3

# Utilities
Pillow==10.3.0
openpyxl==3.1.2
xlwt==1.3.0
reportlab==3.6.13
WeasyPrint==59.0
django-import-export==3.2.0

# Caching & Performance
django-redis==5.2.0
django-debug-toolbar==4.1.0

# Development
pytest==7.3.1
pytest-django==4.5.2
coverage==7.2.7
black==24.3.0
flake8==6.0.0

# Production
gunicorn==23.0.0
whitenoise==6.4.0
sentry-sdk==2.8.0